LayoutTests/inspector/css/generateCSSRuleString.html

<!DOCTYPE html>
<html>
<head>
<style>
*{a-b:1;c-d:2}
body>*{a-b:1;c-d:2}
@media all {body>div{a-b:1;c-d:2}}
@media only screen and (min-width:0px) {body>#test-node{a-b:1;c-d:2}}
@media only screen and (min-width:0px) {@media only screen and (min-height:0px) {body>div#test-node{a-b:1;c-d:2}}}
</style>
<script src="../../http/tests/inspector/resources/inspector-test.js"></script>
<script>
function test() {
    let nodeStyles = null;

    let suite = InspectorTest.createSyncSuite("CSS.generateCSSRuleString");

    suite.addTestCase({
        name: "CSS.generateCSSRuleString.InlineStyle",
        description: "Check the formatting of the generated inline style string.",
        test() {
            InspectorTest.log(nodeStyles.inlineStyle.generateCSSRuleString());
        }
    });

    suite.addTestCase({
        name: "CSS.generateCSSRuleString.MatchedRules",
        description: "Check the formatting of the generated string for all matched CSS rules.",
        test() {
            for (let rule of nodeStyles.matchedRules)
                InspectorTest.log(rule.style.generateCSSRuleString());
        }
    });

    WI.domManager.requestDocument((documentNode) => {
        WI.domManager.querySelector(documentNode.id, "#test-node", async (contentNodeId) => {
            if (!contentNodeId) {
                InspectorTest.log("DOM node not found.");
                InspectorTest.completeTest();
                return;
            }

            let domNode = WI.domManager.nodeForId(contentNodeId);
            nodeStyles = WI.cssManager.stylesForNode(domNode);
            if (nodeStyles.needsRefresh)
                await nodeStyles.awaitEvent(WI.DOMNodeStyles.Event.Refreshed);

            suite.runTestCasesAndFinish();
        });
    });
}
</script>
</head>
<body onload="runTest()">
    <p>Testing that generated CSS rule strings have proper formatting.</p>
    <div id="test-node" style="a-b:1;c-d:2"></div>
</body>
</html>
